# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EmailSecurity API
  version: "1.1"
  description: OpenAPI spec for EmailSecurity backend. Import this into Postman.
servers:
  - url: http://127.0.0.1:8000
    description: Local dev
security:
  - bearerAuth: []
paths:
...
  /policies:
    get:
      summary: List policies
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Policies
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PolicySummary'
  /policy:
    post:
      summary: Create policy (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyCreate'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
  /policy/{id}:
    get:
      summary: Get policy
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Policy content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyDetail'
        '404':
          description: Not found
    put:
      summary: Update policy (Admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyCreate'
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete policy (Admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
  /policy/{id}/activate:
    patch:
      summary: Activate policy (Admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  active:
                    type: boolean
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PolicySummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, inactive]
        last_modified:
          type: string
          format: date-time
    PolicyCreate:
      type: object
      required: [content]
      properties:
        name:
          type: string
          nullable: true
        content:
          type: string
    PolicyDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        active:
          type: boolean
        content:
          type: string
...